Equip pet
Equip pet takes the guts of the mage only arm pet functionality and makes it available to any class with a pet giving them the ability to arm their own pet. Mages still retain the ability to arm other player pets (arm pet should still work the same as before, but shares code with equip pet). Also working on faster method to restock pet gear both on the mage and for other players to request it.