Update e3_ClearXTargets.inc
Adding an option to ForceOn and ForceOff clearxtargets command. 

Currently ClearXTargets is only a toggle, and it's easy to click this multiple times in a frenzied fight and not know what state you're in.   Adding in a ForceOn allows me to mash the button to my hearts content and be assured that I'm engaging mobs.